rag-assistant/requirements.txt
# Core
langchain>=0.2.12
langchain-community>=0.2.10
langchain-text-splitters>=0.2.2
langchain-huggingface>=0.1.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
pydantic>=2.7.0
python-dotenv>=1.0.1
typer>=0.12.3
rich>=13.7.1
gradio>=4.44.0

# Optional LLM providers
openai>=1.40.0
# For local models
ollama>=0.3.0

# Testing
pytest>=8.2.0
